Refactoring and booking feature store structure updated
