--- doctrro-ui/src/app/_modules/booking/_models/search-criteria.spec.ts
@@ -1,7 +1,0 @@
-import { SearchCriteria } from './search-criteria';
-
-describe('SearchCriteria', () => {
-  it('should create an instance', () => {
-    expect(new SearchCriteria()).toBeTruthy();
-  });
-});
