+++ doctrro-ui/src/app/_modules/booking/_models/search-filters.spec.ts
@@ -1,0 +1,7 @@
+import { SearchFilters } from './search-filters';
+
+describe('SearchFilters', () => {
+  it('should create an instance', () => {
+    expect(new SearchFilters()).toBeTruthy();
+  });
+});
