+++ doctrro-ui/src/app/_modules/booking/_services/search.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import {HttpClient} from "@angular/common/http";
-import {DoctorSearchResponse, SearchCriteria} from "../_models";
+import {DoctorSearchResponse, SearchFilters} from "../_models";
 import {Observable} from "rxjs";
 
 const SEARCH_SERVICE = '/search-service/';
@@ -10,7 +10,7 @@ export class SearchService {
 
   constructor(private http: HttpClient) { }
 
-  searchDoctor(criteria: SearchCriteria): Observable<DoctorSearchResponse[]> {
+  searchDoctor(criteria: SearchFilters): Observable<DoctorSearchResponse[]> {
     let searchParams = `lat=${criteria.loc.lat}&lon=${criteria.loc.lon}&radius=${criteria.radius}`;
     if (criteria.speciality) {
       searchParams = `${searchParams}&speciality=${criteria.speciality}`;
