+++ doctrro-ui/src/app/_modules/booking/store/search/filters/search-filter.action.ts
@@ -1,5 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import {SearchCriteria} from "../../_models";
+import {SearchFilters} from "../../../_models";
 
 export const getSearchCriteria = createAction(
   '[SEARCH_CRITERIA] GET'
@@ -7,5 +7,5 @@ export const getSearchCriteria = createAction(
 
 export const updateSearchCriteria = createAction(
   '[SEARCH_CRITERIA] UPDATE',
-  props<{ criteria: SearchCriteria }>()
+  props<{ criteria: SearchFilters }>()
 );
