+++ doctrro-ui/src/app/_modules/booking/store/search/filters/search-filter.reducer.ts
@@ -1,0 +1,13 @@
+import {Action, createReducer, on, State} from "@ngrx/store";
+import * as CriteriaActions from './search-filter.action';
+import {SearchFilters} from "../../../_models";
+
+const searchCriteriaReducer = createReducer(
+  undefined,
+  on(CriteriaActions.getSearchCriteria, state => state),
+  on(CriteriaActions.updateSearchCriteria, (state, { criteria }) => ({ ...criteria }))
+);
+
+export function reducer(state: State<SearchFilters> | undefined, action: Action) {
+  return searchCriteriaReducer(state, action);
+}
