+++ doctrro-ui/src/app/_modules/booking/store/search/search.reducer.ts
@@ -1,13 +1,8 @@
-import {Action, createReducer, on, State} from "@ngrx/store";
-import * as CriteriaActions from './search.action';
-import {SearchCriteria} from "../../_models";
+import {combineReducers} from '@ngrx/store';
+import * as fromSearchCriteria from './filters/search-filter.reducer';
+import * as fromDoctors from './doctors/doctors.reducer';
 
-const searchCriteriaReducer = createReducer(
-  undefined,
-  on(CriteriaActions.getSearchCriteria, state => state),
-  on(CriteriaActions.updateSearchCriteria, (state, { criteria }) => ({ ...criteria }))
-);
-
-export function reducer(state: State<SearchCriteria> | undefined, action: Action) {
-  return searchCriteriaReducer(state, action);
-}
+export const reducer = combineReducers({
+  filters: fromSearchCriteria.reducer,
+  doctors: fromDoctors.reducer
+})
